Delete old header & rename new one
